Rename package to opentracing_shim
